requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib
streamlit
imbalanced-learn
pyyaml